% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_initialize.R
\docType{data}
\name{.rswat_db}
\alias{.rswat_db}
\title{Put this file last in the load order!}
\format{
An object of class \code{rswat_db} of length 1.
}
\usage{
.rswat_db
}
\description{
The code below defines an rswat_db reference class object to store data and information
about the currently loaded SWAT+ project. The class, its methods, and any dependent
functions must be defined first. Make sure that all R files appear in the 'include' tag
above. This ensures the R files are collated in the right order.
}
\keyword{datasets}
