% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_R5.R
\docType{class}
\name{rswat-class}
\alias{rswat-class}
\alias{rswat_generator}
\title{An rswat_path object stores the project directory and executable path}
\description{
swat_dir is the path to the SWAT+ project directory (should contain "file.cio")
exe_path is the path to the SWAT simulation executable (path should end in ".exe")
}

