% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_amatch}
\alias{rswat_amatch}
\title{Approximate assignment based on a distance matrix}
\usage{
rswat_amatch(m)
}
\arguments{
\item{m}{a non-negative numeric distance matrix}
}
\value{
a length-\code{ncol(m)} integer vector mapping labels or \code{NA}
}
\description{
Finds an assignment of \code{nrow(m)} labels to \code{ncol(m)} objects based on
the pairwise distances in \code{m} (where \code{m[i,j]} is the distance between the
\code{i}th label and the \code{j}th object), with no label assigned more than once.
}
\details{
This uses a simple greedy algorithm where in each iteration the least distance
map is assigned and the label removed from the pool of candidates. The function
returns a vector indexing the label for each object. If there are fewer labels
than objects, this vector will have \code{NA}s for unmapped objects.
}
