% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_run.R
\name{rswat_backup}
\alias{rswat_backup}
\title{Save a copy of a SWAT+ project, or a subset of its files}
\usage{
rswat_backup(
  dest = NULL,
  include = NULL,
  exclude = NULL,
  zip = FALSE,
  overwrite = FALSE,
  wipe = FALSE,
  quiet = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{dest}{character, path to the desired output file or folder}

\item{include}{character vector, a set of file, group, or type names to include}

\item{exclude}{character vector, a set of file, group, or type names to exclude}

\item{zip}{logical, indicates to write a zip instead of a folder}

\item{overwrite}{logical, enables modifying or deleting contents of dest}

\item{wipe}{logical, deletes the file or folder dest before writing to it}

\item{quiet}{logical, suppresses console output}

\item{.db}{rswat reference object, for internal use}
}
\value{
character vector, the path(s) to the file(s) written
}
\description{
Makes a copy of some or all of the contents of the currently loaded SWAT+ project
directory. Call the function without arguments to create a new backup of all SWAT+
files known to \code{rswat} in a sub-directory.
}
\details{
Specify other paths for the backup using \code{dest}. Set \code{zip=TRUE} to compress
to a zip file instead, at path \code{dest}. Set \code{overwrite=TRUE} to allow \code{rswat}
to make changes to \code{dest} on disk, and set \code{wipe=TRUE} to first delete any existing
file(s) found at \code{dest}.

By default, the output directory or file is written to the currently loaded SWAT+ project
directory and is named '.rswat_backup_\\<DATE\\>', where \\<DATE\\> is the current date-time
as reported by \code{Sys.time()}, in snake case. If a zip file is being written, the extension
'.zip' is appended automatically (if it isn't there already). Note that on some operating
systems the prefix '.' hides the file unless it is referenced specifically.

Subsets of the SWAT+ project can be be specified by \code{include} and \code{exclude}, as
described in \code{?rswat_files}. When they are both \code{NULL}, the function copies all known
SWAT+ files. Note that this includes 'weather' and 'output' type files, which can be
large in size but which also tend to compress well.

Files unknown to \code{rswat} because they are not listed in 'file.cio' are excluded
automatically (regardless of \code{include} and \code{exclude}), as are existing \code{rswat}
backups and sub-directories.
}
