% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_initialize.R
\name{rswat_check}
\alias{rswat_check}
\title{Check that the rswat package was loaded properly and .rswat_db was found}
\usage{
rswat_check(quiet = FALSE, .db = .rswat_db)
}
\arguments{
\item{quiet}{logical, indicates to suppress console messages}

\item{.db}{rswat_db object, for internal use}
}
\value{
nothing
}
\description{
When the \code{rswat} package is loaded, an object named \code{.rswat_db} is defined and
(by default) passed to function calls through argument \code{.db}. This stores information
about the SWAT+ project as a reference class object that can be accessed globally
(see \code{?rswat_db}). Do not modify (or copy) \code{.rswat_db} unless you know what you are
doing.
}
\details{
This function checks that \code{.rswat_db} exists and is of the correct class. It returns
nothing but has the side effect (when successful) of printing information about
\code{.rswat_db} to the console (unless \code{quiet=TRUE}).
}
\examples{
rswat_check()
}
