% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_run.R
\name{rswat_copy}
\alias{rswat_copy}
\title{Copy SWAT+ config files or make a backup of a project}
\usage{
rswat_copy(
  from = NULL,
  to = NULL,
  fname = NULL,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{from}{character, path to the source directory}

\item{to}{character, path to the destination directory}

\item{fname}{character vector, name of files in \code{from} to copy}

\item{overwrite}{logical, whether to overwrite any existing files in path \code{to}}

\item{quiet}{logical, suppresses console messages}
}
\value{
a vector of paths to the files copied in their new location
}
\description{
Copy SWAT+ config files or make a backup of a project
}
