% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_dir}
\alias{rswat_dir}
\title{Return the currently assigned SWAT+ directory}
\usage{
rswat_dir(.db = .rswat_db)
}
\arguments{
\item{.db}{}
}
\value{
character, the path to the directory or \code{NA} if not yet assigned
}
\description{
If the directory has not been assigned yet, the function returns \code{NA} instead.
}
\details{
This is a wrapper for calling the appropriate method of the \code{.rswat_db} object
created when the package was loaded
}
