% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_docs.R
\name{rswat_docs}
\alias{rswat_docs}
\title{Search variable names and description text in the SWAT+ inputs PDF}
\usage{
rswat_docs(pattern = "*", fuzzy = 0L, defs = FALSE, .db = .rswat_db)
}
\arguments{
\item{pattern}{character vector, the search keyword string}

\item{fuzzy}{integer controlling the number of results returned (see \code{?rswat_fuzzy_search})}

\item{defs}{logical indicating to search definitions text instead of parameter names}

\item{.db}{rswat_db object for internal use}
}
\value{
tibble containing search results and their location in the PDF
}
\description{
Returns a tibble with SWA+ name/definition pairs matching \code{pattern}.
}
\details{
The function uses \code{rswat_fuzzy_search} to look for parameter names containing
the string \code{pattern}. If \code{pattern} exactly matches a SWAT+ file name, the function
returns all known name/definition pairs for that file.
If \code{defs=TRUE} the function searches in definitions instead of names and file
names are ignored. Increase \code{fuzzy} to get approximate results or set it to -1
to get exact matches only (see \code{?rswat_fuzzy_search}).
}
