% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_empty_df}
\alias{rswat_empty_df}
\title{Initialize an empty data frame}
\usage{
rswat_empty_df(nm)
}
\arguments{
\item{nm}{integer or character vector, the length or column names}
}
\value{
data frame with zero rows
}
\description{
Helper function to initialize a data frame of the specified size with zero rows.
If \code{nm} is a character vector, the function returns a data frame with \code{length(nm)}
named columns. If \code{nm} is numeric, the data frame will have \code{nm} unnamed columns.
}
\details{
All output columns are character class.
}
