% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_files}
\alias{rswat_files}
\title{Return a tibble with information about SWAT+ files}
\usage{
rswat_files(
  pattern = NA,
  level = "file",
  include = NULL,
  exclude = NULL,
  known = NA,
  refresh = TRUE,
  quiet = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{pattern}{character, the file name, group, or type of file}

\item{level}{character, either 'file', 'type', or 'group'}

\item{include}{character vector, a set of file, group, or type names to include}

\item{exclude}{character vector, a set of file, group, or type names to exclude}

\item{known}{logical, indicates to only search among recognized files}

\item{refresh}{logical, indicates to first re-scan the directory}

\item{quiet}{logical, suppresses console output}

\item{.db}{rswat_db object, for internal use}
}
\value{
a tibble summarizing the requested files
}
\description{
By default (\code{pattern=NA}) the function returns information on all files in the current
\code{rswat} project directory. Set \code{pattern} to the name of a file, group, or type, to
return information on those files only.
}
\details{
Aggregate results by changing \code{level} from 'file' to 'group' or 'type'.

Filter results by specifying specific file name(s), group(s), or type(s) in
\code{include} and/or \code{exclude}. By default all files are included. Set \code{known=TRUE} to
exclude files unknown to \code{rswat} based on file name(s) and extension(s). Set it to
\code{FALSE} to exclude recognized files and how only unrecognized ones.
}
