% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_find.R
\name{rswat_find}
\alias{rswat_find}
\title{Search SWAT+ config files for a keyword}
\usage{
rswat_find(
  pattern = NULL,
  fuzzy = 0L,
  vals = FALSE,
  add_defs = TRUE,
  quiet = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{pattern}{character vector, the string(s) to search for}

\item{fuzzy}{numeric, specifying tolerance for approximate matches (see details)}

\item{vals}{logical, specifies to search values (in files) instead of parameter names}

\item{add_defs}{logical, appends 'match', 'alias', and 'desc' fields to results}

\item{quiet}{logical, supresses console output}

\item{.db}{rswat_db object, for internal use}
}
\value{
A data frame of information about SWAT+ names matching the search pattern(s)
}
\description{
Searches among loaded SWAT+ files for parameter names matching the search
query \code{pattern}, returning a tibble of information on the matches. If
\code{pattern} is a file name (and \code{vals=FALSE}) the function returns results
on all parameters in the file, automatically loading the file as needed.
}
\details{
The default (\code{fuzzy = 0}) settings return results where \code{pattern} appears as a
sub-string, \code{fuzzy = -1} specifies exact matches only, and  \code{fuzzy > 0} returns,
in addition, the next \code{fuzzy} closest matches. See \code{?rswat_fuzzy_search}.

Definitions are automatically matched to parameters when \code{add_defs=TRUE} (the
default). In cases of discrepancies in spelling, \code{rswat} attempts to match names
automatically to aliases. A three-star match score is assigned, with one or
two stars indicating some doubt.

If \code{vals=TRUE}, the function searches parameter values (as strings) instead of
names and allows file name queries. This can be useful for finding (character
class) parameters that link to other parameters by naming a file.

The special character \code{'*'} matches all parameters in all (loaded) files,
returning everything in a single tibble. This will take a few seconds if you have
\code{add_defs=TRUE}.

Get a list of files and their load state with \code{rswat_open()} or \code{rswat_files()}.
}
