% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_config_methods_helpers.R
\name{rswat_fix_config}
\alias{rswat_fix_config}
\title{Add rows to line_df to represent missing headers in SWAT+ config file}
\usage{
rswat_fix_config(line_df, table_num = 1L)
}
\arguments{
\item{line_df}{data frame of the form returned by \code{rswat_ftable_txt}}

\item{table_num}{integer, the index of the table in the file}
}
\value{
a copy of line_df, possibly with rows added
}
\description{
Add rows to line_df to represent missing headers in SWAT+ config file
}
