% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_config_methods_helpers.R
\name{rswat_ftable_txt}
\alias{rswat_ftable_txt}
\title{Identify tables in SWAT+ config text}
\usage{
rswat_ftable_txt(line_df)
}
\arguments{
\item{line_df}{data frame of the form returned by \code{rswat_scan_txt}}
}
\value{
a copy of \code{line_df} with updated attributes labeling tables
}
\description{
This function splits the line numbers of a plain text file into blocks representing
tables of data. All-character lines are identified as headers, and all subsequent
non-header lines are assumed to belong to the preceding header.
}
\details{
This rule is applied iteratively starting with the first header line. The one exception
is with multiple contiguous lines of all character fields. These are assumed to comprise
one single table. The following columns in line_df are updated:

'tabular' logical, indicates the field is part of a data row in a table
'table' integer, the table number
'skipped' logical, indicating that the line for this field was not assigned to a table
}
