% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_fuzzy_match}
\alias{rswat_fuzzy_match}
\title{Match names to aliases, merging other fields}
\usage{
rswat_fuzzy_match(
  name_df,
  alias_df,
  skip = TRUE,
  alias_desc = TRUE,
  name_split = TRUE,
  alias_split = TRUE,
  div_penalty = 1,
  k_select = NULL
)
}
\arguments{
\item{name_df}{input config file data}

\item{alias_df}{docs data}

\item{skip}{omit from search}

\item{alias_desc}{s}

\item{name_split}{s}

\item{alias_split}{s}

\item{div_penalty}{internal}

\item{k_select}{diagonals}
}
\value{
a
}
\description{
This uses rswat_amatch to match SWAT+ names to aliases in the PDF, then
merges the two data frames. It uses a greedy algorithm for assignment that
favors match patterns that are (nearly) in sequence
}
\details{
Matching is done against the 'name' columns of both inputs
}
