% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_config_methods_helpers.R
\name{rswat_n_prec_txt}
\alias{rswat_n_prec_txt}
\title{Determine precision of numeric columns in a SWAT+ config file}
\usage{
rswat_n_prec_txt(line_df, table_num = NULL, quiet = FALSE)
}
\arguments{
\item{line_df}{data frame of the form returned by \code{rswat_ftable_txt}}

\item{table_num}{integer, the table number to load (or \code{NULL} to load all)}

\item{quiet}{logical, indicates to suppress console messages}
}
\value{
data frame, a copy of \code{line_df} with the 'n_prec' column updated
}
\description{
Determines a the number of digits to appear after the decimal point for numeric types
in columns of a SWAT+ table, based on the existing printed values.
}
\details{
If \code{table_num=NULL}, the function processes all tables.

Most of the time this works without issue on the output of \code{rswat_ftable_txt}, but
there are certain exceptions to deal with. These exceptions are handled by modifying
\code{line_df} after the \code{rswat_ftable_txt} call but prior to this function call (this
happens in \code{rswat} methods ...)
}
