% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_nprec_txt}
\alias{rswat_nprec_txt}
\title{Determine precision of numeric columns in a SWAT+ config file}
\usage{
rswat_nprec_txt(line_df, table_num = NULL)
}
\arguments{
\item{line_df}{data frame of the form returned by \code{rswat_ftable_txt}}
}
\value{
data frame, a copy of \code{line_df} with the 'nprec' column updated
}
\description{
Determines a the number of digits to appear after the decimal point for numeric types
in columns of a SWAT+ table, based on the existing printed values.
}
\details{
If \code{table_num=NULL}, the function processes all tables.

Most of the time this works without issue on the output of \code{rswat_ftable_txt}, but
there are certain exceptions to deal with. These exceptions are handled by modifying
\code{line_df} after the \code{rswat_ftable_txt} call but prior to this function call (this
happens in \code{rswat} methods ...)
}
