% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_run.R
\name{rswat_ohg}
\alias{rswat_ohg}
\title{Configure SWAT+ for daily output to an object hydrograph (OHG) file}
\usage{
rswat_ohg(
  i = 1L,
  obj = "sdc",
  name = "tot",
  revert = FALSE,
  quiet = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{i}{integer, the "object number", an ID code associated with the desired object}

\item{obj}{character, three-character object type code (see details)}

\item{name}{character, one of 'tot', 'sur', 'lat', 'til', 'rhg' (object-dependent)}

\item{revert}{logical, removes OHG output}

\item{quiet}{logical, suppresses console output}

\item{.db}{rswat reference object, for internal use}
}
\value{
either the data frame written to "object.prt" (revert=TRUE), or NULL (revert=TRUE)
}
\description{
Configures the SWAT+ currently loaded SWAT+ project to product OHG output instead of
the output files specified in "print.prt", or (if \code{revert=TRUE}) vice-versa. This
modifies "file.cio" and creates (or deletes) "object.prt".
}
\details{
OHG files contain information on a specific hydrological variable at a specific
location, such as discharge at the main outlet, or recharge from a particular HRU.
Specify an output with \code{i} (ID number), \code{obj} (object type), and \code{name} (variable
of interest). These arguments be supplied as vectors to configure multiple OHG files
at once.

If \code{revert=TRUE}, the function ignores \code{i}, \code{obj}, and \code{name}, deletes the file
"object.prt" if it exists.

The object type codes for \code{obj} are as follows:

HRUs:                   'hru', 'hlt', 'ru'
reservoirs:             'res'
channels:               'cha', 'sdc'
export coefficients:    'exc'
delivery ratio:         'dr'
outlets:                'out'

and the variable names for \code{name} are:

total:                  'tot'
recharge:               'rhg'
surface:                'sur'
lateral:                'lat'
tile:                   'til'
soil moisture by layer: 'sol'
}
