% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_open}
\alias{rswat_open}
\title{Open a SWAT+ project file with rswat}
\usage{
rswat_open(
  f = NULL,
  refresh = TRUE,
  update_stats = TRUE,
  quiet = FALSE,
  output = TRUE,
  .db = .rswat_db
)
}
\arguments{
\item{f}{character vector, the file name(s) to select}

\item{quiet}{logical, suppresses console output and returns results invisibly}

\item{.db}{rswat_db object, for internal use}
}
\value{
a data frame or a list of them, the parameter tables in the config file
}
\description{
Returns data frame corresponding to the data values in the SWAT+ file (or a list of them).
}
