% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_order_columns}
\alias{rswat_order_columns}
\title{Return a subset of data frame columns in a specified order}
\usage{
rswat_order_columns(df_unordered, nm_print)
}
\arguments{
\item{df_unordered}{a data frame with named columns}

\item{nm_print}{column names to return, in order}
}
\value{
a data frame, with (possibly) reordered columns
}
\description{
Columns of \code{df_unordered} named in \code{nm_print} are returned. \code{nm_print} may contain
names not appearing in \code{df_unordered}, but there should be at least one match or else
an empty data frame is returned without warning.
}
