% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_write.R
\name{rswat_prewrite}
\alias{rswat_prewrite}
\title{Prepare a data frame for writing to a SWAT+ file}
\usage{
rswat_prewrite(new_df, quiet = FALSE, .db = .rswat_db)
}
\arguments{
\item{new_df}{data frame, the new SWAT+ parameter table}

\item{quiet}{logical, whether to warn of possible problems with input}

\item{.db}{rswat reference object, for internal use}
}
\value{
data frame, a modified version of \code{new_df} ready to print
}
\description{
an existing SWAT+ parameter table with a new, possibly modified version,
}
\details{
Generate a formatted version of the input data frame, ready to be printed to a
file. For internal use by \code{rswat_write}.

The function replaces missing columns, drops unnecessary ones, coerces mismatched
classes, and converts all numeric columns to character with the same precision level
as found in the file. Most related warnings and messages can be disabled with
\code{quiet=TRUE}.

Logical columns are converted to character, as 'y'/'n'.
}
