% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_rename}
\alias{rswat_rename}
\title{Rename columns in a data frame according to a list of aliases}
\usage{
rswat_rename(d, aliases = NULL)
}
\arguments{
\item{d}{a data frame to check for aliases}

\item{aliases}{named list of character vectors (the aliases)}
}
\value{
the data frame \code{d}, possibly with new names
}
\description{
This checks the names of \code{d} for any matches to the strings in
\code{aliases}. If any matches are found, the column is renamed to the
alias, unless the alias is already in use.
}
