% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_helpers.R
\name{rswat_report_outputs}
\alias{rswat_report_outputs}
\title{Summarize the list of activated output filenames in print.prt}
\usage{
rswat_report_outputs(
  lazy = TRUE,
  n_head = 10L,
  ok_char = .rswat_ok_char(),
  .db = .rswat_db
)
}
\arguments{
\item{lazy}{logical, indicates to load print.prt as needed}

\item{n_head}{integer, the character width to use for printing headers}

\item{ok_char}{character vector of symbols to use for bullets}

\item{.db}{}
}
\value{
a character string to print on the console (with \code{base::cat})
}
\description{
Summarize the list of activated output filenames in print.prt
}
