% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_write.R
\name{rswat_restore}
\alias{rswat_restore}
\title{Restore a backup of a SWAT+ project, or a subset of its files}
\usage{
rswat_restore(
  backup = NULL,
  dest = NULL,
  overwrite = FALSE,
  quiet = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{backup}{character, path to the}

\item{dest}{character, path to the desired output file or folder}

\item{overwrite}{logical, enables modifying or deleting contents destination directory}

\item{quiet}{logical, suppresses console output}

\item{.db}{rswat reference object, for internal use}
}
\value{
character vector, the path(s) to the file(s) written
}
\description{
\code{backup} must point to either a directory with SWAT+ project files, or
else a zip file containing them. If \code{backup} is \code{NULL}, the function returns a tibble
of information on available backups found in the current SWAT+ project directory.
}
\details{
Users must toggle \code{overwrite=TRUE} to enable copying. When it is \code{FALSE} (the default),
the function returns a data frame of files that would be created/modified, but makes
no changes to the files on disk.

The call to \code{utils::unzip} uses \code{junkpaths=TRUE}, so any directory structure in the
zip file will be collapsed (probably a bad thing when unintended). It is recommended
to only load zip files created by \code{rswat} (see \code{?rswat_export})
}
