% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_config_methods_helpers.R
\name{rswat_rtable_txt}
\alias{rswat_rtable_txt}
\title{Extract a table from SWAT+ config text}
\usage{
rswat_rtable_txt(line_df, txt_path, table_num = NULL)
}
\arguments{
\item{line_df}{data frame of the form returned by \code{rswat_ftable_txt}}

\item{txt_path}{character, the path to the SWAT+ text file to be read}

\item{table_num}{integer, the index of the table in the file}
}
\value{
data frame, the specified parameter table in the file
}
\description{
Load a table from a SWAT+ file via \code{data.table::fread} and return it as a data frame.
}
