% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_config_methods_helpers.R
\name{rswat_scan_txt}
\alias{rswat_scan_txt}
\title{Scan data values from a SWAT+ config file line-by-line}
\usage{
rswat_scan_txt(txt = NULL, f = NULL, type = "config", delim = "\\\\s+")
}
\arguments{
\item{txt}{character vector, the output of \code{readLines(f)} (or a subset)}

\item{f}{character, the file name associated with the txt data}

\item{delim}{character, regex for separator between fields (default is any whitespace)}
}
\value{
a data frame of information about the fields in the file
}
\description{
This builds a data frame information about individual fields detected in the file.
These should be delimited by the character \code{delim} (usually whitespace). The output
numbers the fields on each line, detects their types, and copies the field itself as
a string with outer white-space removed.
}
\details{
Other attributes in the output data frame are initialized here but filled in by
other functions (\code{rswat_ftable_txt} and \code{rswat} methods).

If \code{txt} and \code{f} are omitted the function returns a template table with column names
and types assigned, but with 0 rows.

Note that the first line of a SWAT+ config file is assumed to be a comment.
}
