% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_find.R
\name{rswat_show_search}
\alias{rswat_show_search}
\title{Print the results of an rswat search}
\usage{
rswat_show_search(
  results,
  fuzzy = 2L,
  n_max = 10L,
  nm_print = names(results),
  .db = .rswat_db
)
}
\arguments{
\item{results}{data frame with numeric \code{distance} and character \code{name}, \code{file} columns}

\item{fuzzy}{integer, controlling the number of results returned (see \code{?rswat_find})}

\item{n_max}{positive integer, the maximum number of search results to print}

\item{nm_print}{character vector, a subset of \code{names(results)} to return}

\item{.db}{rswat_db object, for internal use}
}
\value{
Returns (invisibly) the subset of \code{results} specified by \code{nm_print}
}
\description{
Helper function for \code{rswat_find} and \code{rswat_docs}. This prints search results to the
console, dividing the results into 1-3 categories (exact, sub-string, or approximate.)
depending on \code{fuzzy}. Categories are based on the match strength scores in column
'distance' (lower is better, see \code{?rswat_fuzzy_search}).
}
\details{
\code{results} should have at least one additional column containing character strings
describing match results. \code{nm_print} controls which columns are printed (normally 'name'
and 'file', and possibly 'description'), and \code{n_max} controls many rows are printed.

The function invisibly returns all rows of \code{results[, nm_print]} (independent of \code{n_max})
}
