% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_helpers.R
\name{rswat_summarize_db}
\alias{rswat_summarize_db}
\title{Helper for show method of rswat_db}
\usage{
rswat_summarize_db(ok_char = .rswat_ok_char(), .db = .rswat_db)
}
\arguments{
\item{ok_char}{character vector of symbols to use for bullets}

\item{.db}{rswat_db object reference, for internal use}
}
\value{
nothing
}
\description{
\code{ok_char} should have names 'yes', 'fail', 'no', and 'sub', corresponding to
items that succeed, fail, have yet to be attempted, and those that are sub-tasks .
}
