% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_validate_dpath}
\alias{rswat_validate_dpath}
\title{Validity check for a directory}
\usage{
rswat_validate_dpath(d, d_name = "d")
}
\arguments{
\item{d}{character, the directory path to check}

\item{vname}{character, name to use in reference to \code{d} (for error messages)}
}
\value{
logical, indicating if the directory has been assigned
}
\description{
Helper function to check if a directory string points to an existing location.
}
\details{
Returns its first argument \code{d} if the directory is found on disk. If the directory doesn't
exist, or if \code{d} is not a valid directory string, the function throws an error. The only
exception is \code{d=NA} (allowing the directory to be unassigned) which returns \code{NA}
}
