% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_validate_fpath}
\alias{rswat_validate_fpath}
\title{Validity check for a file path}
\usage{
rswat_validate_fpath(p, extension = NA, p_name = "p")
}
\arguments{
\item{p}{character, the file path to check}

\item{extension}{character, the file extension to expect including the period (eg '.exe' or '.txt')}

\item{p_name}{character, name to use in reference to \code{p} (for error messages)}
}
\value{
logical, indicating if the directory has been assigned
}
\description{
Helper function to check if a file path string points to an existing file with the given extension.
}
\details{
If a file with the expected extension is found on disk at the path \code{p}, the function returns \code{p}.
Otherwise it throws an error. The only exception is \code{p=NA} (allowing the path to be unassigned),
which returns \code{NA}

Set \code{extension=NA} to allow any file extension.
}
