% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_helpers.R
\name{rswat_weather_report}
\alias{rswat_weather_report}
\title{Return a printable string describing state of climate files in the SWAT+ project}
\usage{
rswat_weather_report(
  lazy = TRUE,
  n_head = 10L,
  quiet = FALSE,
  ok_char = .rswat_ok_char(),
  .db = .rswat_db
)
}
\arguments{
\item{lazy}{logical}

\item{n_head}{integer}

\item{quiet}{logical, if FALSE the function passes its output to \code{base::cat}}

\item{ok_char}{character vector of symbols to use for bullets}

\item{.db}{rswat_db object reference, for internal use}
}
\value{
returns the character string invisibly
}
\description{
Return a printable string describing state of climate files in the SWAT+ project
}
