% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_write.R
\name{rswat_write}
\alias{rswat_write}
\title{Write parameters to SWAT+ config file(s)}
\usage{
rswat_write(
  new_df,
  overwrite = FALSE,
  fast = FALSE,
  quiet = FALSE,
  refresh = TRUE,
  .db = .rswat_db
)
}
\arguments{
\item{new_df}{data frame or list, the SWAT+ table(s) to write}

\item{overwrite}{logical, by default FALSE. Set to TRUE to write changes to disk}

\item{fast}{logical, enables fixed delimiters for faster writing}

\item{quiet}{logical, suppresses console messages}

\item{refresh}{logical, if \code{TRUE} the function reloads the file after modifying it}

\item{.db}{rswat reference object, for internal use}
}
\value{
a tibble listing changes to make in the affected SWAT+ files
}
\description{
Returns a tibble summarizing modifications to SWAT+ configuration file(s) corresponding
to the changes found in the data frame(s) \code{new_df}. When \code{overwrite=TRUE}, these changes
are written to disk.
}
\details{
\code{new_df} should be a data frame returned by \code{rswat_open}, or a list of them. Be aware of
the attributes in these objects with names beginning with 'rswat_'. If some operation
drops the attributes, \code{rswat_write} will not be able to locate the source file.

Set \code{fast=TRUE} to omit most of the summary tibble output and return only the names of the
files to modify. If \code{overwrite=TRUE} and \code{fast=TRUE} the function also renders tables using
\code{data.table::fwrite} (instead of \code{print.data.frame}). This is much faster with large files,
but results in configuration files that are less readable to humans (columns are not aligned).

\code{fast=TRUE} mode is mandatory for 'weather' type input data files ('pcp1.pcp' etc).

Argument \code{refresh=FALSE} disables automatic reloading of files. This is useful for speeding
performance on bulk overwrites of large files (like weather data) but not recommended for
interactive sessions.
}
