% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_zip}
\alias{rswat_zip}
\title{Create zip archives}
\usage{
rswat_zip(zipfile, files, quiet = FALSE)
}
\arguments{
\item{zipfile}{character, the path to the zip file to write}

\item{files}{character vector, file paths to zip}

\item{quiet}{logical, if TRUE nothing is printed to the console}
}
\value{
status value (see utils::zip)
}
\description{
This code is copied from utils::zip and simplified to use default arguments,
but with the new option to omit the console output (with \code{quiet=TRUE})
}
