% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rwat_progress}
\alias{rwat_progress}
\title{Make a vector of messages for reporting progress in console via \code{base::cat}}
\usage{
rwat_progress(nm, n_max = NULL)
}
\arguments{
\item{nm}{character vector, the objects to progress through}

\item{n_max}{integer or NULL, the maximum line width for printing}
}
\value{
a character vector the same length as \code{nm}
}
\description{
This takes a vector of object names and returns a short message for each one,
consisting of a percentage - how far through the list are we - followed by the
name.
}
\details{
Strings are by default truncated/padded to \code{n_max}, the console width as given
by the R option 'width', and each begins with the special symbol '\r', which
(when passed to \code{cat}) asks the terminal or console in which R is running to
do a carriage return and overwrite the existing line.
}
