% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_date_conversion}
\alias{rswat_date_conversion}
\title{Convert between two representations of date}
\usage{
rswat_date_conversion(d)
}
\arguments{
\item{d}{either a Date object or an integer vector or list of them (or a matrix or dataframe)}
}
\value{
Either a Date vector or a data frame of j, yr, values
}
\description{
If \code{d} is a Date vector, the function returns a data frame with the same number of rows,
and attributes: Julian date ('j' ) and year ('yr'). Otherwise the function assumes \code{d}
is a data frame (or matrix, or list) of this form and it does the inverse, returning a vector
of Dates. \code{d} can also be a single integer vector c(j, yr)
}
\examples{
# convert from integers to Dates
n_test = 10
date_as_int = lapply(seq(n_test), \(x) c(j=sample.int(365, 1), yr=1950 + sample.int(100, 1)))
date_result = rswat_date_conversion(date_as_int)
str(date_result)

# and back again
int_result = rswat_date_conversion(date_result)
all.equal(date_as_int, int_result)

# input can be data frame
date_as_df = data.frame(do.call(rbind, date_as_int))
all.equal(date_result, rswat_date_conversion(date_as_df))

}
