% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_config_methods.R
\name{rswat_db_config_methods}
\alias{rswat_db_config_methods}
\title{Methods for opening and parsing SWAT+ config files}
\description{
These define further methods for an rswat_db reference class object, related
to opening configuration files
}
