% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_docs.R
\name{rswat_docs}
\alias{rswat_docs}
\title{Search variable names and description text in the SWAT+ I/O PDF}
\usage{
rswat_docs(name = NULL, f = NULL, quiet = FALSE, .db = .rswat_db)
}
\arguments{
\item{f}{character vector, the file names (sections of document) to include in search}

\item{.db}{rswat_db object, for internal use}

\item{pattern}{character vector, the search keyword string}

\item{fuzzy}{integer, controlling the number of results returned (see details)}

\item{intro}{logical, if TRUE, prints the introductory text for a file}

\item{descriptions}{logical, if TRUE, searches description text instead of variable names}

\item{desc_len}{integer, maximum number description text characters to return (per variable)}
}
\value{
data frame containing search results and their location in the PDF
}
\description{
Search variable names and description text in the SWAT+ I/O PDF
}
