% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_files}
\alias{rswat_files}
\title{Return a data frame of information about SWAT+ files}
\usage{
rswat_files(
  f = NULL,
  what = .rswat_gv_cio_show(),
  type = "config",
  include = "loaded",
  refresh = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{f}{character vector, the file name(s) to select}

\item{what}{character vector, the column(s) to select}

\item{type}{character, one of 'config' (the default)...}

\item{include}{character, one of 'loaded', 'known', 'all'}

\item{refresh}{logical, indicates to re-scan the directory}

\item{.db}{rswat_db object, for internal use}
}
\value{
the requested data frame or vector, a subset of \code{cio}
}
\description{
Convenience function for grabbing subsets of the file information data frame
\code{cio} located at \code{.rswat[['dir']][[swat_dir]][['cio']]}.
}
\details{
Call this without arguments to get a complete copy of the \code{cio} data frame
Specify the column(s) to return with \code{what}. When \code{what} is a single name, the
function returns the requested data in a vector; When \code{what} is a vector, it
returns a data frame. When \code{what=NULL}, all columns are returned.

\code{refresh=TRUE} causes the SWAT+ directory to be scanned for changes
and \code{cio} rebuilt with the latest file information.

\code{f} filters rows by specifying file name(s) (default \code{NULL} selects all),
}
