% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_files}
\alias{rswat_files}
\title{Return a data frame of information about SWAT+ files}
\usage{
rswat_files(
  pattern = NA,
  loaded = NA,
  known = NA,
  include = NULL,
  exclude = NULL,
  what = NULL,
  n = NULL,
  refresh = TRUE,
  summarize = "file",
  quiet = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{pattern}{character, the file name, group, or type of file}

\item{loaded}{logical, indicates to only search among loaded files}

\item{known}{logical, indicates to only search among recognized files}

\item{what}{character vector, the column(s) to select}

\item{n}{integer, the maximum number of results to print}

\item{refresh}{logical, indicates to first re-scan the directory}

\item{quiet}{logical, suppresses console output}

\item{.db}{rswat_db object, for internal use}

\item{type}{character, one of 'config' (the default)...}
}
\value{
the requested data frame or vector, a subset of \code{cio}
}
\description{
If \code{pattern} is the name of a file in the SWAT+ project directory, the function
returns a one-row data frame with information about the file. If \code{pattern} is \code{NA},
information on all files is returned. Otherwise all matches of \code{pattern} against
'group' and 'type' are returned.
}
\details{
\code{loaded} and \code{known} have no effect when set to \code{NA}, or when \code{pattern} is a file
name. Otherwise they control the subset of results to return. \code{loaded=TRUE}
restricts to files loaded already, and \code{loaded=FALSE} restricts to those not loaded.
\code{known=TRUE} restricts results to file types known to \code{rswat} (based on file name
and extension) and \code{known=FALSE} returns only unrecognized files.

\code{what} controls which columns to return. Call \code{.rswat_gv_cio_show()} to see the
defaults, or \code{names(rswat_files(what=NULL))} to see all options.

By default at most \code{n=15} rows are printed for tidiness, but the entire set of
results is returned invisibly. Set \code{quiet=TRUE} to print nothing.
}
