% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_files}
\alias{rswat_files}
\title{Return a data frame of information about SWAT+ files}
\usage{
rswat_files(
  pattern = NA,
  type = NA,
  loaded = NA,
  known = TRUE,
  what = .rswat_gv_cio_show(),
  quiet = FALSE,
  n_max = 15L,
  sort_by = "standard",
  refresh = TRUE,
  .db = .rswat_db
)
}
\arguments{
\item{type}{character, one of 'config' (the default)...}

\item{what}{character vector, the column(s) to select}

\item{refresh}{logical, indicates to re-scan the directory}

\item{.db}{rswat_db object, for internal use}

\item{f}{character vector, the file name(s) to select}

\item{include}{character, one of 'loaded', 'cio', 'known', 'all'}
}
\value{
the requested data frame or vector, a subset of \code{cio}
}
\description{
TODO: details here
}
\details{
\code{f} filters rows by specifying file name(s) (default \code{NULL} selects all),
}
