% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_find}
\alias{rswat_find}
\title{Search SWAT+ config files for a keyword in variable names}
\usage{
rswat_find(
  pattern = "*",
  fuzzy = -1,
  trim = 2,
  n_max = 10L,
  include = rswat_files(what = "file", include = "known", .db = .db),
  .db = .rswat_db
)
}
\arguments{
\item{pattern}{character vector, the string(s) to search for}

\item{fuzzy}{numeric, specifying tolerance for approximate matches (see details)}

\item{trim}{integer between 0-3, higher means more simplification in results}

\item{n_max}{positive integer, the maximum number of search results to return}

\item{include}{character vector, file (or file group) names to include in search}

\item{.db}{rswat_db object, for internal use}
}
\value{
A data frame of information about SWAT+ names matching the search pattern(s)
}
\description{
Returns a data frame of information on matches of \code{pattern} against SWAT+ variable
names. The default \code{'*'} matches everything, resulting in a data frame summarizing
ALL fields in the specified file(s).
}
\details{
\code{fuzzy = -1} is for exact matches only, \code{fuzzy = 0} includes substring matches, and
\code{fuzzy > 0} includes approximate substring matches. See \code{?rswat_string_distance}.

By default the function searches all loaded files. Otherwise, the function searches
in the subset of files specified by \code{include}, loading missing files as needed.
Get a list of files and their load state with \code{rswat_open()}.
}
