% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_find.R
\name{rswat_find}
\alias{rswat_find}
\title{Search SWAT+ config files for a keyword}
\usage{
rswat_find(
  pattern = "*",
  f = .db$get_loaded_files(),
  fuzzy = 2L,
  trim = 3L,
  n_max = 10L,
  quiet = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{pattern}{character vector, the string(s) to search for}

\item{f}{character vector, file (or file group) names to include in search}

\item{fuzzy}{numeric, specifying tolerance for approximate matches (see details)}

\item{trim}{integer between 0-3, higher means more simplification in printed result}

\item{n_max}{positive integer, the maximum number of search results to return}

\item{quiet}{logical, supresses console output}

\item{.db}{rswat_db object, for internal use}
}
\value{
A data frame of information about SWAT+ names matching the search pattern(s)
}
\description{
Returns a data frame of information on matches of \code{pattern} against SWAT+ file and
variable names. The default \code{'*'} matches everything, resulting in a data frame
summarizing ALL fields in the specified file(s).
}
\details{
\code{fuzzy = -1} is for exact matches only, \code{fuzzy = 0} includes results where
\code{pattern} appears as a sub-string, and  \code{fuzzy > 0} includes the next \code{fuzzy} closest
matches. See \code{?rswat_fuzzy_search}.

By default the function searches all loaded files. Otherwise, the function searches
in the subset of files specified by \code{include}, loading missing files as needed.
Get a list of files and their load state with \code{rswat_open()}.
}
