% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_docs.R
\name{rswat_match_docs}
\alias{rswat_match_docs}
\title{Match SWAT+ variable names found in config files to definitions in documentation}
\usage{
rswat_match_docs(file_df, quiet = FALSE, .db = .rswat_db)
}
\arguments{
\item{file_df}{input config file data}

\item{quiet}{logical, supresses console output}

\item{.db}{rswat_db object, for internal use}
}
\value{
a tibble with columns alias, match, definition, appended
}
\description{
Match SWAT+ variable names found in config files to definitions in documentation
}
