% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_open}
\alias{rswat_open}
\title{Open a SWAT+ project file with rswat}
\usage{
rswat_open(f = NULL, quiet = FALSE, .db = .rswat_db)
}
\arguments{
\item{f}{character vector, the file name(s) to select}

\item{.db}{rswat_db object, for internal use}
}
\value{
a data frame or a list of them
}
\description{
Returns data frame, or a list of them, corresponding to the data values
found in the SWAT+ file.
}
