% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_config_methods_helpers.R
\name{rswat_scan_dir}
\alias{rswat_scan_dir}
\title{Scan a SWAT+ project directory}
\usage{
rswat_scan_dir(swat_dir = NULL, cio_df = NULL)
}
\arguments{
\item{swat_dir}{path to the directory containing SWAT+ config files}

\item{cio_df}{optional data frame structured like the output of \code{rswat_scan_dir()}}
}
\value{
a data frame with a list of file info for \code{swat_dir}
}
\description{
Looks for files in the supplied \code{swat_dir} and builds a data frame listing them
along with some OS file info (size, etc) and internal flags. Files are labelled
according to known SWAT+ file extension patterns.
}
\details{
An existing data frame, \code{cio_df}, can be supplied to list files not found on disk
alongside the ones detected by \code{base::list.files}
}
