% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_find.R
\name{rswat_show_search}
\alias{rswat_show_search}
\title{Print the results of an rswat search}
\usage{
rswat_show_search(
  results,
  fuzzy = 2L,
  n_max = 10L,
  nm_print = names(results),
  quiet = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{results}{data frame with numeric \code{distance} and character \code{name} columns}

\item{fuzzy}{integer, controlling the number of results returned (see \code{?rswat_find})}

\item{n_max}{positive integer, the maximum number of search results to print}

\item{nm_print}{character vector, usually a subset of \code{names(results)} (see details)}

\item{.db}{rswat_db object, for internal use}
}
\value{
Returns (invisibly) the subset of \code{results} specified by \code{nm_print}
}
\description{
Helper function for \code{rswat_find} and \code{rswat_docs}. This prints search results to the
console, dividing the results into 1-3 categories (exact, sub-string, or approximate.)
depending on \code{fuzzy}. Categories are based on the match strength scores in column
'distance' (lower is better, see \code{?rswat_fuzzy_search}).
}
\details{
\code{results} must have an additional column named 'nm_lu', containing the character
strings that were matched. \code{nm_print} controls which columns are printed (normally 'name'
and 'file', and possibly 'description'), and \code{n_max} controls many rows are printed.

If a name in \code{nm_print} is not found among the columns of \code{results}, the function
uses the \code{get_cio_df} method of \code{.db} to look it up. Get a list of valid names for
this feature with \code{names(.rswat_db$get_cio_df())}.

All rows of \code{results[, nm_print]} are returned invisibly, independent of \code{n_max}.
}
