% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_summarize_files}
\alias{rswat_summarize_files}
\title{Print a summary of a data frame of SWAT+ file information}
\usage{
rswat_summarize_files(files_df, level = "file", quiet = FALSE, .db = .rswat_db)
}
\arguments{
\item{files_df}{data frame, output from \code{rswat_files}}

\item{level}{character, the category to summarize over (see details)}

\item{quiet}{logical, suppresses console messages}
}
\value{
returns a data frame invisibly, and prints the first \code{n} lines of it
}
\description{
A helper function for \code{rswat_files} to print file information in the console.
\code{level} specifies the detail level for the output. The default \code{level='file'}
returns information about every file on its own row.
}
\details{
With \code{level='group'} each group gets a row, indicating how many files there are
and how many are loaded, along with a list of the file names. \code{level='type'} does
the same but with summaries of the four file types, and group names within each type
category instead of file names. In both cases, the time modified field indicates the
latest modification within the category
}
