% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat.R
\name{rswat_time}
\alias{rswat_time}
\title{Get or set the SWAT+ simulation time parameters}
\usage{
rswat_time(sim = NULL, prt = NULL, step = NULL, .db = .rswat_db)
}
\description{
WORK IN PROGRESS
}
\details{
Helper function for reading or setting up parameters SWAT+ parameters
controlling the time steps executed by the simulator, as well as the
requested time steps (and variables) to print to output files.

These parameters are located in two files:
\itemize{
\item "time.sim" controls the simulation start date, end date, and step size
\item "print.prt" controls which output variables get printed
}
}
