% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_run.R
\name{rswat_time}
\alias{rswat_time}
\title{Configure simulation time for a SWAT+ simulation}
\usage{
rswat_time()
}
\arguments{
\item{dates}{integer, vector, or data frame indicating the dates to simulate}

\item{nyskip}{integer, the number of years to skip in printed output}
}
\value{
something
}
\description{
This modifies 'print.prt' and 'time.sim' on disk to configure a simulation over the
requested dates. Outputs are printed starting \code{nyskip} years after the starting date.
}
