% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_utils.R
\name{rswat_truncate_txt}
\alias{rswat_truncate_txt}
\title{Render table data to the specified character width for printing}
\usage{
rswat_truncate_txt(
  txt,
  max_len = NULL,
  NA_char = "~",
  more_char = "...",
  just = "left"
)
}
\arguments{
\item{txt}{character vector, or a data frame}

\item{max_len}{integer, the maximum width allowed for printing (\code{NULL} to set automatically)}

\item{NA_char}{character, string to print in place of \code{NA}s}

\item{more_char}{character, string to indicate that a field has been shortened}
}
\value{
character vector or data frame (the same length as \code{txt})
}
\description{
Render table data to the specified character width for printing
}
