% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_db_helpers.R
\name{rswat_weather_dates}
\alias{rswat_weather_dates}
\title{Fetch the range of dates in a SWAT+ weather input file}
\usage{
rswat_weather_dates(lazy = TRUE, .db = .rswat_db)
}
\arguments{
\item{lazy}{logical, indicates to load 'weather-sta.cli' and all of the files listed therein}

\item{.db}{rswat_db object reference, for internal use}
}
\value{
returns a data frame of start and end dates
}
\description{
work in progress
}
