% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rswat_write.R
\name{rswat_write}
\alias{rswat_write}
\title{Write parameters to SWAT+ config file(s)}
\usage{
rswat_write(
  new_df,
  overwrite = FALSE,
  fast = FALSE,
  quiet = FALSE,
  .db = .rswat_db
)
}
\arguments{
\item{new_df}{data frame or list, the SWAT+ table(s) to write}

\item{overwrite}{logical, by default FALSE. Set to TRUE to write changes to disk}

\item{fast}{logical, enables fixed delimiters for faster writing}

\item{quiet}{logical, suppresses console messages}

\item{.db}{rswat reference object, for internal use}
}
\value{
a tibble listing changes to make in the affected SWAT+ files
}
\description{
Returns a tibble summarizing modifications to SWAT+ configuration file(s) corresponding
to the changes found in the data frame(s) \code{new_df}. When \code{overwrite=TRUE}, these changes
are written to disk.
}
\details{
\code{new_df} should be a data frame returned by \code{rswat_open}, or a list of them.

Set \code{fast=TRUE} in combination with \code{overwrite=TRUE} to omit the summary tibble output
and render tables using \code{data.table::fwrite}. This is much faster with large files, but
results in less readable configuration files (columns are not aligned).
}
